Pause / resume history + drawing tool refactoring
