+++ examples/nextjs-whiteboard/src/Ellipse.tsx
