+++ examples/nextjs-whiteboard/src/IconButton.tsx
