+++ examples/nextjs-whiteboard/src/Rectangle.tsx
