+++ examples/nextjs-whiteboard/src/SelectionBox.module.css
