+++ examples/nextjs-whiteboard/src/SelectionBox.tsx
