+++ examples/nextjs-whiteboard/src/SelectionTools.module.css
