+++ examples/nextjs-whiteboard/src/SelectionTools.tsx
