+++ examples/nextjs-whiteboard/src/hooks.ts
@@ -1,4 +1,5 @@
 import { LiveMap, LiveObject } from "@liveblocks/client";
+import { useRoom } from "@liveblocks/react";
 import { useState, useEffect } from "react";
 import { Layer, XYWH } from "./types";
 import { boundingBox } from "./utils";
@@ -8,22 +9,19 @@ export function useSelectionBounds(
   selection: string[]
 ): XYWH | null {
   const [bounds, setBounds] = useState(boundingBox(layers, selection));
+  const room = useRoom();
 
   useEffect(() => {
-    onChange();
-
     function onChange() {
       setBounds(boundingBox(layers, selection));
     }
 
+    onChange();
+
     // We need to subscribe to the layers map updates and to the updates on the layers themselves.
     // If User A deletes or modified a layer that is currently selected by UserB, the selection bounds needs to be refreshed.
-    layers.subscribeDeep(onChange);
-
-    return () => {
-      layers.unsubscribeDeep(onChange);
-    };
-  }, [layers, selection]);
+    return room.subscribe(layers, onChange, { isDeep: true });
+  }, [room, layers, selection]);
 
   return bounds;
 }
