+++ examples/nextjs-whiteboard/src/types.ts
@@ -51,7 +51,7 @@ export type Presence = {
   selection: string[];
   cursor: Point | null;
   // TODO: Fix Serializable type to support array of array
-  penPoints: any[] | null;
+  pencilDraft: any[] | null;
   penColor: Color | null;
 };
 
@@ -88,7 +88,7 @@ export type CanvasState =
       current: Point;
     }
   | {
-      mode: CanvasMode.Drawing;
+      mode: CanvasMode.Inserting;
       layerType: LayerType.Ellipse | LayerType.Rectangle;
     }
   | {
@@ -105,11 +105,32 @@ export type CanvasState =
     };
 
 export enum CanvasMode {
+  /**
+   * Default canvas mode. Nothing is happening.
+   */
   None,
+  /**
+   * When the user's pointer is pressed
+   */
   Pressing,
+  /**
+   * When the user is selecting multiple layers at once
+   */
   SelectionNet,
+  /**
+   * When the user is moving layers
+   */
   Translating,
-  Drawing,
+  /**
+   * When the user is going to insert a Rectangle or an Ellipse
+   */
+  Inserting,
+  /**
+   * When the user is resizing a layer
+   */
   Resizing,
+  /**
+   * When the pencil is activated
+   */
   Pencil,
 }
