+++ examples/nextjs-whiteboard/src/useDisableScrollBounce.module.css
@@ -1,0 +1,4 @@
+.no_scroll {
+  overscroll-behavior: none;
+  overflow: hidden;
+}
